split mwnu and catalogue files into different folders
